move helper function out of hooks folder
